social_media_api/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-taggit==6.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.0
whitenoise==6.9.0
